+++ fs-helpers.js
